Switch the language hint (if it even matters)
I don’t expect that GitHub’s markdown supports the relatively obscure RouterOS scripting language, but it does show the language percentage on the front page, so something in their system is aware of it.

The search link on that percentage points to a `routeros-script` language, so I figured I’d make my hints match that in case it matters at all.